fix(wise): ImageRegion shift int cast for slice indices
ImageRegion.set_shift stored np.round(shift) as float64; downstream
get_region_slice() built slice(-(x0+dx), None) etc., and Py3 strictly
requires int slice indices. Cast the rounded result to int.

+++ packages/libwise/src/libwise/imgutils.py
@@ -1703,7 +1703,7 @@ class ImageRegion(Image):
     def set_shift(self, shift):
         ''' shift shall be of type int or will be rounded '''
         if shift is not None:
-            self.shift = np.round(shift)
+            self.shift = np.round(shift).astype(int)
         else:
             self.shift = None
